+++ src/main/java/io/github/notsyncing/lightfur/utils/PageUtils.java
@@ -2,7 +2,7 @@ package io.github.notsyncing.lightfur.utils;
 
 public class PageUtils
 {
-    public static int calculatePageCount(int pageSize, int totalCount)
+    public static int calculatePageCount(int pageSize, long totalCount)
     {
         if (totalCount == 0) {
             return 0;
@@ -16,7 +16,7 @@ public class PageUtils
             return 1;
         }
 
-        int pageCount = totalCount / pageSize;
+        int pageCount = (int)(totalCount / pageSize);
 
         if (totalCount % pageSize > 0) {
             pageCount++;
